# sdk-trace: use `MapRef` in `SpanStorage`