--- a/sdk/all/src/test/scala/org/typelevel/otel4s/sdk/OpenTelemetrySdkSuite.scala
+++ b/sdk/all/src/test/scala/org/typelevel/otel4s/sdk/OpenTelemetrySdkSuite.scala
@@ -401,9 +401,7 @@ class OpenTelemetrySdkSuite extends CatsEffectSuite {
       s"OpenTelemetrySdk{meterProvider=$meterProvider, " +
       "tracerProvider=" +
       s"SdkTracerProvider{resource=$resource, spanLimits=$spanLimits, sampler=$sampler, " +
-      "spanProcessor=SpanProcessor.Multi(" +
-      s"BatchSpanProcessor{exporter=$exporter, scheduleDelay=5 seconds, exporterTimeout=30 seconds, maxQueueSize=2048, maxExportBatchSize=512}, " +
-      "SpanStorage)}, " +
+      s"spanProcessor=BatchSpanProcessor{exporter=$exporter, scheduleDelay=5 seconds, exporterTimeout=30 seconds, maxQueueSize=2048, maxExportBatchSize=512}}, " +
       s"propagators=$propagators}, resource=$resource}"
 
 }