--- a/sdk/trace/src/test/scala/org/typelevel/otel4s/sdk/trace/SdkTracesSuite.scala
+++ b/sdk/trace/src/test/scala/org/typelevel/otel4s/sdk/trace/SdkTracesSuite.scala
@@ -292,9 +292,7 @@ class SdkTracesSuite extends CatsEffectSuite {
   ) =
     "SdkTraces{tracerProvider=" +
       s"SdkTracerProvider{resource=$resource, spanLimits=$spanLimits, sampler=$sampler, " +
-      "spanProcessor=SpanProcessor.Multi(" +
-      s"BatchSpanProcessor{exporter=$exporter, scheduleDelay=5 seconds, exporterTimeout=30 seconds, maxQueueSize=2048, maxExportBatchSize=512}, " +
-      "SpanStorage)}, " +
+      s"spanProcessor=BatchSpanProcessor{exporter=$exporter, scheduleDelay=5 seconds, exporterTimeout=30 seconds, maxQueueSize=2048, maxExportBatchSize=512}}, " +
       s"propagators=$propagators}"
 
 }