## Benchmarks

Before:
```
Benchmark                               (backend)   Mode  Cnt      Score      Error   Units
TraceBenchmark.span                           sdk  thrpt   80  37936.668 ± 1867.799   ops/s
TraceBenchmark.span:async                     sdk  thrpt             NaN                ---
TraceBenchmark.span:gc.alloc.rate             sdk  thrpt   80   1721.229 ±   62.352  MB/sec
TraceBenchmark.span:gc.alloc.rate.norm        sdk  thrpt   80  48424.089 ± 2845.034    B/op
TraceBenchmark.span:gc.count                  sdk  thrpt   80     93.000             counts
TraceBenchmark.span:gc.time                   sdk  thrpt   80   5132.000                 ms
```

After:
```
Benchmark                               (backend)   Mode  Cnt      Score      Error   Units
TraceBenchmark.span                           sdk  thrpt   80  43930.485 ± 1372.618   ops/s
TraceBenchmark.span:async                     sdk  thrpt             NaN                ---
TraceBenchmark.span:gc.alloc.rate             sdk  thrpt   80   1698.339 ±   59.193  MB/sec
TraceBenchmark.span:gc.alloc.rate.norm        sdk  thrpt   80  40544.982 ±  530.282    B/op
TraceBenchmark.span:gc.count                  sdk  thrpt   80     85.000             counts
TraceBenchmark.span:gc.time                   sdk  thrpt   80   5635.000                 ms
```

There is ~16% improvement in throughput and ~17% less memory allocation. 

## oteljava vs SDK

```
Benchmark                  (backend)   Mode  Cnt      Score      Error   Units
TraceBenchmark.noopScope        noop  thrpt   80  194900.259 ± 4233.790   ops/s
TraceBenchmark.noopScope    oteljava  thrpt   80  167703.033 ± 2820.815   ops/s
TraceBenchmark.noopScope         sdk  thrpt   80  130521.459 ± 5903.623   ops/s
TraceBenchmark.pure             noop  thrpt   80  203382.916 ± 4371.169   ops/s
TraceBenchmark.pure         oteljava  thrpt   80  194281.902 ± 3699.690   ops/s
TraceBenchmark.pure              sdk  thrpt   80  187178.501 ± 5383.391   ops/s
TraceBenchmark.rootScope        noop  thrpt   80  204148.742 ± 3868.173   ops/s
TraceBenchmark.rootScope    oteljava  thrpt   80  103563.256 ± 7134.662   ops/s
TraceBenchmark.rootScope         sdk  thrpt   80   44412.737 ±  952.992   ops/s
TraceBenchmark.span             noop  thrpt   80  202511.661 ± 4672.375   ops/s
TraceBenchmark.span         oteljava  thrpt   80  113092.272 ± 7984.096   ops/s
TraceBenchmark.span              sdk  thrpt   80   45129.609 ± 1034.677   ops/s
```